requirements.txt
python-telegram-bot[socks]==21.0.1